stack-structs: drop redundant usize cast on element_size
Clippy on the validation spot flagged `*element_size as usize` as a same-type
cast — `ArrayAccess::element_size` is already `usize`. Lint-only fix; no
behaviour change. Suite was already green on the prior spot run.

+++ crates/hexray-analysis/src/decompiler/stack_struct_binding.rs
@@ -626,12 +626,9 @@ fn transform_expr(
                 {
                     if let Some(binding) = bindings.containing(byte_off) {
                         let rel = (byte_off - binding.stack_offset) as usize;
-                        if let Some(field_expr) = field_access_for_struct_offset(
-                            binding,
-                            rel,
-                            *element_size as usize,
-                            db,
-                        ) {
+                        if let Some(field_expr) =
+                            field_access_for_struct_offset(binding, rel, *element_size, db)
+                        {
                             return field_expr;
                         }
                     }
